Refactor schedules into a single module
After using Dagster for a while I realize that I never need more than a
single python module to contain all of my Dagster schedules.
